Update vendored simdjson to v4.6.4
- Download simdjson.h and simdjson.cpp from v4.6.4 release
- Update SIMDJSON_VERSION with new hashes
- Fix .golangci.yml for golangci-lint v2 (add version field, remove
  deprecated gosimple linter)
- Fix unsafe pointer conversions flagged by govet unsafeptr check

Closes #4

+++ ndjson.go
@@ -37,8 +37,8 @@ func ParseND(b []byte, reuse *ParsedJson, opts ...ParserOption) (*ParsedJson, er
 		return nil, fmt.Errorf("%s", C.GoString(res.result.error_msg))
 	}
 	pj.tape = Tape{
-		data:        copyUint64Slice(uintptr(unsafe.Pointer(res.tape)), int(res.tape_len)),
-		strings:     copyByteSlice(uintptr(unsafe.Pointer(res.sbuf)), int(res.sbuf_len)),
+		data:        copyUint64Slice(unsafe.Pointer(res.tape), int(res.tape_len)),
+		strings:     copyByteSlice(unsafe.Pointer(res.sbuf), int(res.sbuf_len)),
 		copyStrings: pj.copyStrings,
 	}
 	pj.hasTape = true
